requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.2
django-dotenv==1.4.2
django-environ==0.11.2
django-rest-swagger==2.2.0
django-uwsgi==0.2.2
djangorestframework==3.14.0
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
openapi-codec==1.3.2
packaging==23.2
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
